gcc 13 mingw fixes
  * changes needed for mingw build

+++ src/support/serial.cxx
@@ -780,7 +780,7 @@ int  Cserial::ReadBuffer (std::string &buf, int nchars, std::string find1, std::
 	snprintf(traceinfo, sizeof(traceinfo), 
 		"ReadBuffer [%0.2f msec] (%u): %s",
 		readtime, sbuf.length(), 
-		(hex ? str2hex(buf.c_str(), buf.length()) : sbuf.c_str()));
+		(hex ? str2hex(buf.c_str(), (unsigned  int)buf.length()) : sbuf.c_str()));
 	if (progStatus.serialtrace)
 		ser_trace(2, "1:", traceinfo);
